+++ src/DataSource.ts
@@ -10,7 +10,7 @@ import {
   TimeRange,
   toDuration,
 } from '@grafana/data';
-import { DataSourceWithBackend, getTemplateSrv, TemplateSrv } from '@grafana/runtime';
+import { DataSourceWithBackend, getTemplateSrv, TemplateSrv, config } from '@grafana/runtime';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
@@ -61,11 +61,11 @@ export class XrayDataSource extends DataSourceWithBackend<XrayQuery, XrayJsonDat
       const params = new URLSearchParams({ region });
       searchString = '?' + params.toString();
     }
-    return this.getResource(`/groups${searchString}`);
+    return this.getResource(`groups${searchString}`);
   }
 
   async getRegions(): Promise<Region[]> {
-    const response = await this.getResource('/regions');
+    const response = await this.getResource('regions');
     return [
       ...sortBy(
         response.map((name: string) => ({
@@ -76,6 +76,22 @@ export class XrayDataSource extends DataSourceWithBackend<XrayQuery, XrayJsonDat
         'label'
       ),
     ];
+  }
+
+  async getAccountIdsForServiceMap(range?: TimeRange, group?: Group): Promise<string[]> {
+    if (!config.featureToggles.cloudWatchCrossAccountQuerying) {
+      return [];
+    }
+    const params = new URLSearchParams({
+      startTime: range ? range.from.toISOString() : '',
+      endTime: range ? range.to.toISOString() : '',
+      group: group?.GroupName || 'Default',
+    });
+
+    const searchString = '?' + params.toString();
+
+    const response = await this.getResource(`accounts${searchString}`);
+    return response.map((account: { Id: string }) => account.Id);
   }
 
   getServiceMapUrl(region?: string): string {