+++ src/utils/transform.ts
@@ -375,7 +375,6 @@ export function parseGraphResponse(response: DataFrame, query?: XrayQuery, optio
   for (const service of services) {
     const statsSource = service.SummaryStatistics ? service : service.Edges[0];
     const stats = statsSource.SummaryStatistics;
-
     idField.values.add(service.ReferenceId);
     titleField.values.add(service.Name);
     typeField.values.add(service.Type);
@@ -404,6 +403,11 @@ export function parseGraphResponse(response: DataFrame, query?: XrayQuery, optio
   for (const edgeData of edges) {
     const { edge, source } = edgeData;
     const target = servicesMap[edge.ReferenceId];
+    // when filtering results by account id, not every target/source will always be returned
+    // if this is the case, there's no need to render an edge
+    if (!target) {
+      continue;
+    }
     edgeIdField.values.add(source.ReferenceId + '__' + target.ReferenceId);
     edgeSourceField.values.add(source.ReferenceId);
     edgeTargetField.values.add(edge.ReferenceId);
@@ -477,26 +481,44 @@ export function parseGraphResponse(response: DataFrame, query?: XrayQuery, optio
 }
 
 export function avgResponseTime(statistics: SummaryStatistics) {
+  if (!statistics.TotalResponseTime || !statistics.TotalCount) {
+    return 0;
+  }
   return (statistics.TotalResponseTime / statistics.TotalCount) * 1000;
 }
 
 export function tracesPerMinute(statistics: SummaryStatistics, startTime: number | string, endTime: number | string) {
+  if (!statistics.TotalCount) {
+    return undefined;
+  }
   return endTime && startTime ? statistics.TotalCount / ((toMs(endTime) - toMs(startTime)) / (60 * 1000)) : undefined;
 }
 
 export function successPercentage(statistics: SummaryStatistics) {
+  if (!statistics.OkCount || !statistics.TotalCount) {
+    return 0;
+  }
   return statistics.OkCount / statistics.TotalCount;
 }
 
 export function throttledPercentage(statistics: SummaryStatistics) {
+  if (!statistics.ErrorStatistics || !statistics.TotalCount) {
+    return 0;
+  }
   return statistics.ErrorStatistics.ThrottleCount / statistics.TotalCount;
 }
 
 export function errorsPercentage(statistics: SummaryStatistics) {
+  if (!statistics.ErrorStatistics || !statistics.TotalCount) {
+    return 0;
+  }
   return (statistics.ErrorStatistics.TotalCount - statistics.ErrorStatistics.ThrottleCount) / statistics.TotalCount;
 }
 
 export function faultsPercentage(statistics: SummaryStatistics) {
+  if (!statistics.FaultStatistics || !statistics.TotalCount) {
+    return 0;
+  }
   return statistics.FaultStatistics.TotalCount / statistics.TotalCount;
 }