+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.3.0
+- Feature: Make it possible to filter on account id in https://github.com/grafana/x-ray-datasource/pull/157
+
+## 2.2.0
+- Make properties of `SummaryStatistics` optional
+
 ## 2.1.2
 
 - Security: Upgrade Go in build process to 1.19.3